+++ test_unstructured_inference/inference/test_layout_element.py
@@ -1,5 +1,5 @@
-from layoutparser.elements import TextBlock
-from layoutparser.elements.layout_elements import Rectangle as LPRectangle
+
+
 
 from unstructured_inference.constants import Source
 from unstructured_inference.inference.layoutelement import LayoutElement, TextRegion
@@ -24,22 +24,3 @@ def test_layout_element_from_region(mock_rectangle):
     assert LayoutElement.from_region(region) == expected
 
 
-def test_layout_element_from_lp_textblock():
-    mock_text_block = TextBlock(
-        block=LPRectangle(100, 100, 300, 300),
-        text="Sample Text",
-        type="Text",
-        score=0.99,
-    )
-
-    expected = LayoutElement.from_coords(
-        100,
-        100,
-        300,
-        300,
-        text="Sample Text",
-        source=Source.DETECTRON2_LP,
-        type="Text",
-        prob=0.99,
-    )
-    assert LayoutElement.from_lp_textblock(mock_text_block) == expected